+++ src/ekf/ekf.h
@@ -19,7 +19,7 @@ public:
     void init(Vec3 * gyro, Vec3 * acc, Vec3 * mag);
     void update_acc(Vec3 acc);
     void update_mag(Vec3 mag);
-    void predict(Vec3 gyro, double dt);
+    void predict(Vec3 gyro, float dt);
     void update();
 };
